feat: implements token refresh flow and simplifies token requests
Implements token_refresh command to perform a token refresh. Token refresh expects a refresh token as input, either provided with a command-line parameter or through stdin, while the output is in the same format as other commands.

This also simplifies the implementation of the token request and brings it in line with the authorization request implementation in terms of using gorilla/schema for constructing the request based on the provided attributes.