feat: adds flag for introspection response format (#53)
Adds a --response-format flag to the introspect command. This flag can be used to indicated the desired format of the response, in particular to choose between json and a JWT. The default value is json. The value required to retrieve a JWT is dependent on the implementation of the authorization server that is used (e.g. for Ping Federate it is token-introspection+jwt while for Keycloak it is just jwt).